Only calculate new views when appropriate.
We did pretty large updates to the model in ModelMod, preventing us from
accurately tidying the graph at the right times. By getting rid of
ModelMod, and only updating the backingNodes (new name for savedNodes)
we can be more precise in recalculating the view nodes (aka tidying the
graph), meaning we don't need to call tidyGraph everywhere just in
case (which broke dragging).