Fix reentering node behaviour introduced by last bug.
Originally, when we reentered a node, afterwards we would create a
ResultHole, which was weird when changing nodes already in the graph.

We changed the behaviour to not do that, but went too far: if the node
had a child, we would ignore ParamHoles too.

Now, we'll only Select if it's a ResultHole