Fix another bug in the reentry condition.
Previously I fixed nodes with children, when we were editing their
params. When we finish editing the param, we want to select the node,
not open a ResultHole.

However, when you press enter on a node, you do want to create a
ResultHole, regardless of whether it had children. The problem was we
routed both situations through the same path. Now we have a new path for
refocus.