Remove Pointers.
Anything that used a pointer is replaced with an ID or PointerData.

Anything that's used immediately is probably using a PointerData, anything
that's related to model state is probably using an ID.

This removes a class of bugs where we got whether something was PFilled or
PBlank wrong. More importantly, it stops us using both Pointers and IDs in weird
ways, or requiring a pointer where the only information we have is an ID.

There's more cleanup to do, but this is enough to ship. In the future, planning
to rename PointerData to Pointer.