Fix the "browser crashed" appearance when scrolling the canvas.
The browser previously stopped moving (giving the appearance of being crashed),
because we update the position via the Navigation module (which updates the url
hash and then tells us about it and lets us respond). When scrolling, we were
sending so many location/navigation updates that the browser gave up.

Instead, we update the app with SetCenter, and update the browser/URL in the
background. I tried many other solutions (see
https://trello.com/c/fYMAPadr/584-scrolling-or-holding-the-arrow-key-down-both-stop-after-a-while)
and this was the most reliable and simple.