Many changes to Dockerfile, getting it to work.
The primary problem was that there was a bug in nocrypto that led to it changing
it's dependencies on opam (it can do that apparently). It not depends on sexplib
<0.11, which of course brings all of core to 0.10 instead of 0.11, despite what
we told opam.

Why was it able to do that? Because when we specify one-line-at-a-time opam
installs, opam fixes things so you can do the latest thing. So I changed it to
install in one line, which made the resolver act up. I changed to opam 2.0 so
I'd get better errors and a better resolver.

Got some problems building re2 that involved missing the build-essential
package.

While I was there, I pulled all the "quick hacks" inline, and restructured the
file to be more sectional.

Discovered a problem with our locale-gen code - no idea why it worked before.

Also, we were totally using lwt 3.1.0 before, despite our best efforts.

We ended up on core 0.10, where I suspect we'll stay until nocrypto releases
0.5.5.

OK, I think that's all.