Create a way to deal with a parse error in json.
Dealing with the string "<Binary Data>" which is invalid json. Rather than try
to figure out all the possible ways that json can be invalid, let's give the
user the ability to deal with it. If parsing fails, response.body will have an
error in it, but response.raw will have the original string, which they can fix
up before parsing with JSON::read.