Make updating the canvas really fast (60fps), even for dabblefox-shelf.
This caches all html and only rebuilds the tl that the cursor is in.

Typing, scrolling, etc, are all instant now.

We might need to do some cache invalidation. Perhaps we should store the
timestamp of analyses so that we know if they have been changed.

The biggest issue I can imagine here is that DBs locked status might be incorrect.