Split most of the code into server-only and maybe-client.
Libexecution is the client name.

I toyed with multiple ways of splitting it, and found that two libraries was
simplest.

This also makes the compilation really fast using --dev.

I've disabled a bunch of stuff, all marked as TODO SPLIT. I'll be reenabling
them as I figure out how to.

This compiles and a number of very simple things still work on the server. The
client isn't close to working yet. I'm working to get the server working first
and then we can figure out the client.