Start better logging.
I want to log parameters but not too many and just the name not the whole sql. I
settled on logging the sql in a debug and the rest in an info. I discovered
there's a better postgres-ocaml interface for this, and that it also has a
binary interface that's 3x faster than what I was doing for saving oplists.