Killed TypeError
We have missed lots of exceptions from TypeError, that we were swallowing in
raise_arglist_error.

Change the fail fn to do the old TypeError check, as well as accepting a msg for
errors within stdlib functions.

Discovered a real bug, which was being dropped. If you run List.filter on a
function which returns incomplete, you get an exception which got
dropped. Instead, you now get an incomplete.

Disabled some API stuff as it's not used anymore.