It has become far too difficult to keep toplevel_oplists and json_oplists in sync, so remove json_oplists.
We're going to be have to be careful when changing the structure of the AST.

When loading from "backup" (which means disk in the case of tests or the
json_oplists db in the case of non-tests), we saved the entire set of oplists in
the DB.

Unfortunately, if we first fetched the backup from a query that only fetched a
subset of toplevels, we'd be in trouble as we'd then only backup a subset of
toplevels.

Solving this, and especially dealing with the consistency issues, was getting
extremely complex, so I decided to kill it instead. Hope that was the right
decision.

This fixes the bug in the rename_db_* integration tests