Fix the DB focus
When adding a new line in a DB toplevel, we go to it using FocusNext. However,
FocusNext relies on having a correct set of toplevels, which we did not have
because we just appended the new toplevels to the end instead of overwriting the
existing ones. So we were looking at the old toplevels, finding the old DB
definition, and then finding the next focus in that one.