Add sexp converters for Uuidm.t
I was running down a bug in the tests in the previous commit
and the error seemed to be in sexp_conv. Turned out it was just
indicating that in the backtrace because all tests using our Sexp
DSL will have that, but in the process of debugging I removed the
failwith's and added basic converters to to our uuid type which
might be useful going forward.