Fix function renaming not propagating to calls
Problem: Users reported that when they renamed a function
that the change did not propagate to the calls of that function.

Actual problem: Renaming worked fine, users were deleting with
backspace before editing as opposed to using 'enter' to start
editing.

Fix: Change `Selection.delete` behaviour for FnName's to bring a
user into Entering with an ACQuery of "", as opposed to deleting
the text with SetFunction. Bonus: hitting escape after backspace
on a fnname no longer breaks everything by having a function with
no name.