Fix Entry showing twice in feature flags
When a feature flag is being shown, we show either a or b twice, once in the
feature flag box, and once in the AST. When we show b twice, since b isn't
created yet, we'll have two entry boxes.

This solves it by only showing the entry box in the feature flag itself.

This has the downside that if you double-click on the ast node, it will open the
entry box in the feature flag, which I think is fine?