Don't allow the server/_build directory to sync
Our build is slow on docker partially because of the amount of syncs in the
build directory, which is really slow because docker syncs it to the host.

There are a number of solutions to this, including:

- mounting an app2 directory and using unison to sync to a separate build
  directory (complex)
- using docker-sync for a similar purpose (complex)
- using docker's "cached" or "delagated" consistency mount flags (helpful, but
  not much).

However, this is really simple, and really effective.

I initially tried this using tmpfs instead, and it didn't have any additional
benefit but required more RAM so I removed it.