The build wasn't working because the `dark` container didnt exist.
Use a known public container instead.

It worked often because CircleCI routes the build to a machine that has already
run our builds for efficiency, and that machine had a dark image on it. When a
second build ran, it would run on another machine which did not have that image.