Show param names in blankOr fields that are filled in
When considering why we show param names in multiline functions, its so the user does not loose sight of where they are and what they my be dealing with.
Given that, we should show param names once the expr is filled regardless if the function params are horizontally or vertically laid out.