Cleanup old traces until yesterday.
As the amount of traffic increases, the amount in the last day will be huge.

Originally, I meant to delete traces more than an hour old, with an hour chosen
so that we wouldn't have any clock skew issues. However, I made it 36
hours when I was testing and didn't move it back. Then I realized I can use the
DB time and avoid clock skew altogether.