Show animation for incoming input value.
So for some odd reason, I can't get the first element to re-render on in the list to render no matter what I try.
Forcing diffs, I also tried getting the whole list to attempt to animate left when it gets new input vals... didn't work either.

After several hours of banging my head against the monitor, I thought we might try this hack of a hack solution.
the "last child" does seem to be a thing that gets re-drawn. So I added a last-child selector to attach the animation keyframes to.
Then I reversed the list in elm, but since we don't want to show the list to show in reverse order, I used flexbox ordering to reverse the list again.

SUPER DUPER HACKY... with lots of not not logics.

Since I felt bad about making something so hacky, I unhackified the rendering input values code, took it out of the attrs and gave it it's own space.
Putting things in attrs gives me an anxiety attack, I worry that the string my be too long or have html breaking characters and the whole thing will mess up rendering.