Fix TL dragging
It seems subscription changes happen only when model gets update, adding the dragging subscription is too late, and the Mouse.move subscription does not get added to our app.

within DragToplevel case handler, it only activates when cursorState is Dragging, it should be safe to register the Mouse.move sub at other points of cursor state, except when they are super irrelevant, such as command entering or entering in code.