On my laptop, fswatch was working, but is no more. Forcing -m poll_monitor fixes it.
Specifically, I was getting events for IsDir and PlatformSpecific, and
nothing else, which is not helpful. Not sure why, and also need to
confirm if this is just my personal laptop or also my dark laptop. (Both
ubuntu 18.04.)

In theory, inotify (default on linux) has worse performance than
poll_monitor; in practice, for a 28k file repo, this should be fine.
http://emcrisostomo.github.io/fswatch/doc/1.5.1/html/fswatch/Monitors.html#The-Poll-Monitor:
"The authors’ experience indicates that fswatch requires approximately
150 MB of RAM memory to observe a hierarchy of 500,000 files with a
minimum path length of 32 characters. A common bottleneck of the poll
monitor is disk access, since stat()-ing a great number of files may
take a huge amount of time."