Add SelectEnterVariant
make the "Select" modification go to Enter state instead

Possible enhancement:
---------------------
Right now, if the exp is on, and I click on a label, it replaces it with an empty box.

If the exp is off, and I have a label "foo" selected and press Enter, I get "foo|" (| being the cursor).

The latter behavior seems more desirable; makes it a bit more obvious
what the pre-Enter state was.

Not sure why the discrepancy. In either case, I see updateMod get called
with `Types.Enter (Filling (tlid, id))` (confirmed by logging).