Fix field_access integration tests.
In Elm, when we typed "req." (to do request.body), after pressing '.', we received

EntryInputMsg "req."
GlobalKeyPress "."

However, in TEA we get
GlobalKeyPress "."
EntryInputMsg "."

Furthermore, since the '.' is part of the input box, further EntryInputMsgs also
container the '.' at the front: ".body" for example.

In Elm, we had a big hack to make it work. I tried a bunch of hacks for TEA and this one
worked for me.