Use strings for ID and TLID on bucklescript
We have an int size mismatch between client + server, and between
elm and bucklescript, and between bucklescript and historical data.

There's no actual reason why we need to use ints for ID/TLID anyway,
they're purely an identifier that we just need a comparison function
for. We should move to using strings everywhere for them as that's
the most portable, and this change will allow us to move to stringified
UUIDs as the identifiers also.