Click the fncall selector without a filter
The text filter is unnecessary as there's only one fncall
on the canvas -- and on the off chance that we introduce
a bug with more than one, the expectation on the generated
uuid live values more than protects us.

The text filter was flakey as ['Uuid', '::', 'generate'] are
the contents of separate spans in the same div and testcafe
seems to sometimes see them as a whole or sometimes separately.
I couldn't quite figure out why but regardless, it's unneccessary.