Merge new traces with old traces
Model.traces is a map of `tlid -> [trace]`, UpdateTraces
is also a map of `tlid -> [trace]` -- but it might not contain the full
set of tlids that we have previously seen in this session because we
only received a subset from the server. We should merge our old set
with the new set, preserving traces for toplevels not in the latest
batch but overwriting all traces for toplevels in the new batch.

This should ensure we have available variables and data for a toplevel
when we click away and come back. This seems to work in Elm without
this patch, but it seems to me like that is accidental and that this
patch correctly expresses our intent.

Warning: I've observed that this brings back our Aw, Snap! behaviour
on linux -- but I think that's because we are now requesting analysis
for all toplevels we have traces for -- as opposed to just the one
we just received back. This extra compute seems to mean the web
workers get created faster than they complete and are GC'd