Cache the routing table's html.
We cache the html for all toplevels via the Utils cache. When we created this,
it was expensive to create a perfect cache key because of how Elm did
things, and the routingTable needed a perfect cache.

However, now that we have refs, it's pretty cheap to use a perfect cache key to
cache the routing table.

This removes about 15ms from each render (in Elm this only took 6ms, so it was
less valuable to focus on).