Canvases don't share editor state.
Prior to this commit, editor state was stored in local storage with the
key "editorState".

In the before times, this worked just fine because each canvas lived
under a separate subdomain. With all canvases living under the same
subdomain (darklang) this doesn't work. "editorState" holds information
about the last canvas that was open. If a developer switches canvases,
this state is meaningless; the application may behave strangely.

Now, key the editor state in local storage using the canvas name.