Remove speed restrictions in integration tests
Prior to this, there were two integration tests that failed if run at
full speed, but would pass if run at a reduced speed.

The failing tests try to press the "Finish integration tests" button
from the function space. At full speed, it seems, the click handler is
not yet registered on the finish button causing the test to time out.
This might indicate a performance issue in the function space, but for
now...

This commit changes the two tests to return to the handler space before
finishing the tests.

It also changes the after hook for each test to:
- More clearly message the failure
- Ensure that the console logs are flushed even when test completion times out