When we run the container in the background, we can timeout. Fix it.
For example, in https://circleci.com/gh/darklang/dark/3328#tests/containers/3,
there was no prebuilt container available, so it built the container on
demand. Because this was done in the background, the "wait-for-container" script
timedout after 5m, failing the build.

The correct thing to do is run the build in full and then run it in the background.