Remove old gcp-local setup
In the old days, we had a single docker image that everything ran on. Back then,
I tried to make a local version of it to run integration tests on. I never got
it to work.

If we were to do local integration tests in a prod-like environment, this
wouldn't be it, so no point keeping it around.