Allow the cursor to appear in function names.
This takes a really hacky approach. In order to put the cursor in an fncall
name, it must be a blankOr. However, doing that on the backend would require a
migration. So instead, the encoders and decoders convert the string name into a
blankors, using a special ID (we append "_name" to the ID).

As a result it feels like you can change a fncall but if you try it errors
weirdly.

To get this over the line for the test:
- we need a better error when you try to change a fncall name
- the positioning is off