Make query_params consistent between the 3 places that use them
Rather than using /integration-test, /ui-debug, ?debug and ?debug=1, as well as
VariantTesting parsing, have a single place where this stuff is parsed on the
frontend and the backend.

This also adds the ~on parameter to String.split on the frontend.

All this in service of a later commit to allow production to use local assets