Place test-GET icon next to lock. & refactor code
Moving icon:
The test-GET icon looked misaligned directly next to get. Moved it to be near lock icon, so we can have a group of actionable icons on the right.

Refactor:
Moved all single use function definitions inside a view header function. Scoped functions locally so they are only exposed in a need to know basis