Put dynamic autocomplete at the bottom.
They're mostly used when there are no other options. However, is a few cases
there are other options and we want those options to come first. Eg, if you're
typing "HT" in an event space, you're likely looking at "HTTP", and it's not
valuable to make you type the whole thing