Remove delete TL logic from Select.delete function
You may also notice that also remove, the error message associated with action.
That was also intentional. The only reason why someone would attempt to backspace on an entire function when they are in functionspace is because they can also do that with DBs and handlers. However if we remove the ability to do that in canvas view, then the association of deleting a thing by hitting DEL is also gone.