This fixes a huge memory leak.
On every autocomplete reset, the matcher would close over the model, including
the old matcher, thus recursively building up a complete history of models and
autocomplete since the dawn of time. This was bad.

This stops the closing over, allowing the memory to free up. This brings chase's
memory usage from "intolerable" to a svelte 200MB or so.

Learning: do not close over the model and then store it on the model.