Remove allocation + dynamic dispatch per req
We previously returned a boxed `Future`, ie. a trait object, on the
heap as `handle` could return _two_ different concrete implementations
of `Future`, `Ok` and the large combinator type of `handled` (Which is
`OrElse<Map<Fold....>>>>>>`). This boxed `Future` gets allocated on the
heap, and more importantly the calls to `poll` on it by the Hyper
executor must be dynamically dispatched via a vtable.

By switching to wrapping these in `Futures::Either`, we only ever return
a single concrete implementation of `Future` (that is the one provided by
`Either`), so we switch to the `impl Future` shorthand to regain
static dispatch and lose an allocation.

For the curious reader, the actual concrete type which is hiding
behind `impl Future` that we return is probably something like

```
Futures::Either<Future::Ok<Item = Response, Error = hyper::Error>,
Future::OrElse<Future::Map::Future::Fold...>>>>>>`
```

Which shows just how useful `-> impl Trait` syntax is!