Fix decoder failure for incoming 404s
This is actually a 4-tuple, not a triple -- so the decoder is failing
and thus the Msg callback is not firing (and so 404s don't live-update).

The silent failure of the decoder is bad, but I'm unsure where's best to
log/crash for this? @pbiggar: any ideas? These could fire _a lot_ for
say, old clients, so I worry about Rollbar here.

Fixes: https://trello.com/c/I9ey4AfR/678-404s-dont-arrive-until-page-refresh