Preserve source ip so nginx' remote_addr in logs is useful in honeycomb
(We want the _user_ IP, not some LB or k8s proxy ip. See
https://trello.com/c/UzNhvX6D/717-nginx-realip-config)

There is a caveat, but this seems acceptable:
https://kubernetes.io/docs/tasks/access-application-cluster/create-external-load-balancer/#caveats-and-limitations-when-preserving-source-ips
```
GCE/AWS load balancers do not provide weights for their target pools. This was not an issue with the old LB kube-proxy rules which would correctly balance across all endpoints.

With the new functionality, the external traffic will not be equally load balanced across pods, but rather equally balanced at the node level (because GCE/AWS and other external LB implementations do not have the ability for specifying the weight per node, they balance equally across all target nodes, disregarding the number of pods on each node).

We can, however, state that for NumServicePods << NumNodes or NumServicePods >> NumNodes, a fairly close-to-equal distribution will be seen, even without weights.
```

(We do have that state - currently 6 nodes and 36+ bwd pods.)