Mostly a revert of #742, but also an update of log_format to match x-dark-username (#743)
Related: https://trello.com/c/qGpTop75/727-put-username-in-response-headers-for-honeycomb-consumption

Turns out, log_format is _only_ for putting the actual log_format string
in, and _can't_ be used to pull in an nginx.conf and get it from that.
(In contrast to running the honeytail cli, which has flags for this.)
Honeycomb has opened a ticket to ... either fix this discrepancy or
document how to use it.