Blow up if we have overlapping same specificity handlers
This whole thing, including the previous iteration could be a
single 'group by'-esque list iteration, bucketing them into
a hashtable -- but we'd need to key directly on the length of the
split path segments. that'd make this algorithm less generic across
how we determine the specificity ordering. unsure what to do, but will
leave it as is rn in case we change what we want our specificity
relation to be