Originally, is_pageable was true or null
But we can't construct a query including a field honeycomb has never
seen, so ... we can't create a trigger/alert for is_pageable=true until
we get a pagable exn. Oops.

By defaulting to false, we get this field added to the schema as soon as
rollbar goes off at all, which means we can put an alert on it quickly.