Make migration script handle relations
Types we don't recognize:
- if they're lists, make them lists of id
- if they're not lists, make them id

I'm not sure this'll play well once we have user-defined types in DBs,
but we don't otherwise have a tag in DB types for TBelongsTo/THasMany.

Additionally: previous round of migrations misinterpreted the cols
field in CreateDBMigration; we only want to change the tipe, not the
name.