Allow using \n, \t, etc in strings
The major issue was that we didn't know when a string was in a particular
format. There are two string formats, a "display" string, and a literal
string. When using autocomplete, and editing in general, we use a display
string. A display string can have \n in it, and also can have just a plain
backslash. A literal string must be valid - it will have bare backslashes
removed.

This commit basically carefully manages strings as they go from
literals (Values) to display (viewable and editable) strings.