Route presence.darklang.com -> sydney-presence.builtwithdark.com
This should allow us to share cookies between the editor and the
presence service.

Other steps to be taken:
- CNAME presence.darklang.com -> darklang.com

To test, post-merge (and post-dns-setup): I've got a / handler up on
sydney-presence; hitting presence.darklang.com/ should result in a trace
with my cookie in the headers.

Part of https://trello.com/c/s5jcFfaD/1005-be-push-data-from-db-to-a-canvas-channel-via-stroller