Comment out scripts/gcp-build-containers step in build-backend
- This keeps failing non-deterministically (sometimes works, sometimes
doesn't) with the error `^@^@cp: cannot stat 'backend/_build/default/bin/server.exe': No such file or directory`
- Example failure: https://circleci.com/gh/darklang/dark/12293
- Leaving this step in build-stroller because we haven't seen it fail
there yet

The intent of this step was that if our ldd check fails (wrong dylibs
available in prod containers, relative to what we compile with), we'd
know pre-merge. With this commit, we'll still be protected against bad
deploys (because the post-merge gcp-build-containers step will fail),
but we will have a broken master branch. These failures (ldd check) are
less common than the missing-exe error that we're seeing in
build-backend, so that's acceptable.

Trello'd investigation/fix/uncommenting of this step at
https://trello.com/c/1GmycHBK/1024-fix-pre-merge-gcp-build-containers-step