Don't center on single click (use double click to center)
This removes centering from single clicks by no longer updating the URL in a way
that triggered the URL handler. It now uses pushState without issuing a
notification to the navigation handler.

We also decided not to institute a nice algorithm for whether to center, as
these have gone substantially wrong on the past, and we know of significant edge
cases (eg huge handlers, either vertically or horizontally).

I also moved page-related stuff to a new file: Page.ml. It wasn't really URL
related so it didnt belong in Url.ml. Url.ml is now restricted to stuff touching
Urls directly.