Make autocomplete work properly
The problem with autocomplete was the each of targetTL, targetDval, query, and
targetTI were different and could be updated independently. This means that we
were updating them independently at different points during an update, sometimes
catching them at an intermediate state which ended up persisting with weird
behaviour.

The most visible manifestation was setting when variables weren't yet known
because we had set targetTL, but not targetTI. This made the index pick a
non-variable, when it could have picked the variable.

The fix is to completely remove state and only pass state in when generating
autocomplete. (We keep the previous query just to compare the old index during
refilter).

As a result, the autocomplete is regenerated with no state, there's no
targetDval being updated at weird times, no need to remember to reset, etc. The
code ends up being way simpler.