Set cookie so it can be shared by subdomains
We want to be able to use our darklang.com cookie on
presence.darklang.com

https://trello.com/c/IGct6S13/1064-set-cookies-we-can-share-with-subdomains

Per
https://www.mxsasha.eu/blog/2014/03/04/definitive-guide-to-cookie-domains/,
"When not setting an explicit domain for a cookie, the default in most
browsers is to only send the cookie when the domain matches exactly. "
(The exception is IE, which we don't care about.)

So we do need to set an explicit domain.  Per RFC6265, it is no longer
necessary to supply a leading dot ('.darklang.com') to trigger this
behavior; 'darklang.com' suffices.

Note: this will not change the behavior of existing cookies! We'll want,
once this is deployed, to drop sessions. This should be invisible to
users not currently in-browser; since we're using basic auth, their
browser should re-auth on request to /a/<canvas>. Users currently
in-browser will have to reload.