Replace occurences on change to variable binding
Modify the `EVariable` and `EPartial` case in `Fluid.replaceTokenString`
when inside a let binding or lambda LHS to search for all
occurences using `Fluid.getVariableOccurences`, which is a slightly
modified version of `AST.uses` from pre-fluid code, and change
each occurence as well. Another special case here for `EVariable` and
`EPartial` handles when a changed occurence becomes a partial and
preserves it as a variable.