Do etags substitution in appsupport.js at build time
https://trello.com/c/2oWvxMp2/1068-make-changes-to-appsupportjs-to-load-from-static

When IanC and I first implemented our static assets, we decided to do
this etags substitution in the webserver, where ocaml was already
running, instead of having to create yet another process in the prod
build.

However, now that we want to load appsupport.js as a file (instead of
just cat'ing it into ui.html), we need to have etag substitution done at
build time, so that the appsupport.js file knows where to find the
serviceworker.js files.

This commit does that.

(Note: we already do this next bit, see hash_static_filenames, so I'm
just leaving this for posterity as a thing I thought was necessary but
is not: We also need to have the appsupport.js url be etag-replaced
if-and-only-if it is being loaded from non-ngrok. (That is, if loading
from ngrok, then we want to load appsupport.js itself; in the normal
non-ngrok case, we want to load appsupport-someetag.js.))