Make the push via stroller contain all the current avatars for a canvas
Instead of multiple pushes per event. This way, we don't need to
dedupe-on-insertion.

Doing this means pushStrollerEvent's payload is now any dval, not just
DObj, bc we want to push a DList of DObjs.

https://trello.com/c/tUEfXfJC/1087-fix-deduping-of-code-in-presence-tlid-avatars

To be done post-deploy:
In dark, instead of ::foreach pushStrollerEvent over the avatars, just
push the whole list as a single event.