When loading a page with an invalid hash url, don't error.
There was error checking logic but it checked in Url.changeLocation, which is
for responding to location changes in the loaded app. This moves the logic to
SetPage where it can handle both changes and the initial load.

It's easy to add an error here if we want one, but after trying it it didn't
seem to add anything.