Created execution state type
It wasn't possible to derive information on whether the handler finished execution from existing information. We needed to know if a handler has just finished triggered execution and thus created a sum type exeState to keep track of : Has not run, running, and finished running.

put that in handler props for now. We really should do a refactor sometime where client data/computed values is for a particular handler is not scattered across the model. Where we constantly have to do filtering

TODO: clean up executingHandlers usage in next commit