Server side: Filter ops if an add_op request comes in out of order
Some of our ops, applied out of order, will succeed but with undesired
effects. Example, a SetHandler that sets values:
"aaaa"
"aaa"
"aa"
"a"

being processed by the server out of order might result in that handler
going:
"aaaa"
"aaa"
"a"
"aa", which is not the desired end state, and also causes a visible
flicker.

So, if an incoming add_op's op_ctr is not the largest seen so far (for
its browser_id), then we should not process its SetHandler op(s).

Other ops, we do not want to drop - for instance, DeleteDBCol must be
run; later ops will not contain state indicating "this DB no longer has
that column".

Client side:
Just like add_op requests can come in to the server out of order,
addOpStrollerMsgs can come in from pusher out of order.

Ergo, keep a dict of <browserId -> opCtr> in the model, and filter
accordingly.

This does _not_ solve potential race conditions between an InitialLoad
and incoming stroller msgs.