Handle max last_op_ctr state
Our bottleneck here is the postgres int, which is 4 bytes, and (since we
start at zero) maxes at 2**31-1. We could use bigint, but then the
bottleneck is 53 bits via encoding, so let's just deal with this on the
server.

This commit adds:
- a log_annotation, "op_ctr", to all add_op calls, so we can see how
that counter grows in honeycomb
- a Rollbar when we reach (2**31 - 1)*0.8

I don't expect this to happen any time soon. (Most add_op calls this
week: ~700, on ellen-dates. Granted, fluid will result in a lot more
such calls, as will multi-user canvases, but ... 1M/day is still 2.5+
years.)