Enter at the end of a match or thread should add a new row
We do this by adding TNewlines to the end of match and thread
expressions.

We also want these final-TNewlines to be invisible, so we don't have
places with an empty div, so TNewline now has a 3rd tuple element
indicating whether its class should be newline-collapsed, in which case
css has it as display:none.

In toTestText, we render these as "~", because we don't allow
zero-length tokens, so something has to be there.