Filter TLSavepoint ops
This is re: a bug whereby stroller messages seemed to not be getting
properly filtered. (Specifically, deleting a chunk of a string, then
reloading, would result in visible replaying of old state.)

What was happening was, we didn't filter out TLSavepoint ops from
"older" messages, and so the tlid in the TLSavepoint in the _params_
meant that we kept that tlid present in result.handlers (instead of
filtering it).

The fix is to add TLSavepoint to the list of ops we ignore if the
message being processed is "old".