Handle unauth response to API endpoints, don't make them redirect
- Ideally, RPC calls wouldn't follow redirects at all. They shouldn't
wihtout an option to. Unfortunately, that option is not exposed to us in
Tea_http. It also seems like jsdom (a dep of jest) might force
followAllRedirects:true

- So - we check in webserver.ml if path starts with /api/, and if it
does, respond 401 instead of redirecting

- We also cause our RPC error handlers to detect and ApiError if this
has happened (err in some cases is a string. not the original error, so
we have to string match instead of status code matching, boo. Which
means we reload if status = 401 _or_ the reload flag is set in
HandleApiError).

- Finally, having an array of Tea_tasks didn't work, so I consolidated
the two Tea_tasks to one.

If we get an error back from trace_results or db_stats, and it contains "Bad credentials", reload