Add honeycomb logs to queue worker
Not strictly required for the other trellos, but improves visibility.

For any log in a qw execution, we now know which canvas, which worker,
etc and the # of retries.

Had to do a little reordering, and use Result.bind instead of just
catching exn's, so we could start a thread containing as much as
possible of the execution (which is how our log annotations work)