Add timer to qw
The approach is a little weird (set a 'start_timer' pair in the
thread-local-variable we use for logs, and then use it when we call
Log.{infO,erroR,etc}) but it lets us do timers across function
boundaries, as long as we're careful not to set start_timers inside each
other.

Querying honeycomb will then be something like, 'get all logs named
'Exception while dequeueing' or 'Successful execution' or whatever your
log is named, and then select the "timer" field and P90 or AVG or
whatever on that'.

Real traces would be nice here, but I think that has to wait for later,
it's not easy to add.

https://trello.com/c/UpcyOri8/1770-add-a-timer-around-queue-and-cron-worker-execution-and-log-runtime-to-honeycomb