Oops this is two squashed commits
Commit 1: Add in enqueued_at and last_processed_at fields to events table

For operational purposes, not product features

This gets us honeycomb logs for:
- time from enqueue to dequeue ('queue_delay_ms')
(We already had dequeue to done processing via in-process timers.)

https://trello.com/c/UpcyOri8/1770-add-a-timer-around-queue-and-cron-worker-execution-and-log-runtime-to-honeycomb

Commit 2:
logs for cron

We sort of get cron logs/timers from the qw stuff, because cron _just_
says "are we ready to run yet? if so, enqueue an event!"

The easy thing to get, which we add here, is the delay in cron itself,
from the scheduled time to the time we enqueued the event.

We do _not_ get the delay from enqueue to dequeue, but that is done by
logging in event_queue.ml.

We _do not_ anywhere get "from scheduled cron time to done processing";
I am assuming that the cron delay is small and so can be ignored, but we
do log cron delay by itself so we'll see if that is wrong.