We have a TLS cert that is not darklang or bwd (hellobirb); let's automate copying it to a new cluster
Esp since we're about to add a similar thing for kiksht.

Motivation: if I run gke-create-cluster without this, the bwd-ingress
can't come up, because it depends on secret/www.hellobirb.com-tls, which
is missing because we didn't copy it over.

I decided _not_ to make bwd-tls and darklang-tls use this function,
because I want those to continue to be overridable by cli flags.