A single bad cron interval should not break all of cron_checker
should_execute's behavior _should_ be (and now is) to just return false
if the interval is unparseable. This allows other crons to be checked,
rather than throwing an exception.

Longer-term, we'll want to validate on save, but this at least keeps all
of cron from breaking for now.

I'm attaching to
https://trello.com/c/7v0WHuqN/1855-prevent-invalid-cron-intervals , but
that's really the follow up ticket.