Move to MOUNTS til after loading the workdir.
I've observed for a few months that directories like _build, stroller/target,
node_modules, and other build directories were found on my local machine instead
of being entirely contained in docker volumes. I suspect the order of the -w and
the -v are important, but I couldn't find anything online to support it.